Refactor deregister method (#274)
